% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{add - staple it together}
\usage{
add(x, y, z = 0)
}
\arguments{
\item{x}{first element}

\item{y}{the next one}

\item{z}{yet another one}
}
\value{
additive result
}
\description{
add - staple it together
}
\examples{
add(2,2)
}
